add ObjectMapper support for Hibernate entities, add documentation
